# Restore performance of recording sealed subclasses