--- a/test/testdata/resolver/sealed_concrete_parent_class.rb
+++ b/test/testdata/resolver/sealed_concrete_parent_class.rb
@@ -23,4 +23,4 @@ def foo(x)
   end
 end
 
-T.reveal_type(ConcreteParent.sealed_subclasses) # error: Revealed type: `T::Set[T.any(T.class_of(Child2), T.class_of(Child3), T.class_of(Child1))]`
+T.reveal_type(ConcreteParent.sealed_subclasses) # error: Revealed type: `T::Set[T.any(T.class_of(Child1), T.class_of(Child2), T.class_of(Child3))]`