--- a/test/testdata/resolver/sealed_module.rb
+++ b/test/testdata/resolver/sealed_module.rb
@@ -21,4 +21,4 @@ def foo(x)
   end
 end
 
-T.reveal_type(Parent.sealed_subclasses) # error: Revealed type: `T::Set[T.any(T.class_of(Child2), T.class_of(Child3), T.class_of(Child1))]`
+T.reveal_type(Parent.sealed_subclasses) # error: Revealed type: `T::Set[T.any(T.class_of(Child1), T.class_of(Child2), T.class_of(Child3))]`